+++ R/utils.R
@@ -74,3 +74,15 @@ utctime <- function(year, month, day, hour, min, sec, psec) {
     as.integer(hour), as.integer(min), as.integer(sec), as.numeric(psec)
   )
 }
+
+cli_block <- function(expr, class = NULL) {
+  msg <- ""
+  withCallingHandlers(
+    expr,
+    message = function(x) {
+      msg <<- paste0(msg, x$message)
+      invokeRestart("muffleMessage")
+    }
+  )
+  rlang::inform(msg, class = class)
+}